Kernel - Prevent MountEntry node access when concurrently removed
Fixes #1191

By having a lock on the MountEntry that is acquired while holding the global lock, we can assure the caller that the object will be alive until he releases it. The delete becomes safer as it can only happen when the MountEntry is locked and now one can get or wait for it due to the global lock.